+++ client/src/Edit/EditPage.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { Container } from "react-bootstrap";
-import BookmarkGroupModal from "../Modal/BookmarkGroupModal";
 import BookmarkModal from "../Modal/BookmarkModal";
 import ConfirmModal from "../Modal/ConfirmModal";
+import GroupModal from "../Modal/GroupModal";
 import Body from "./Body";
 import Header from "./Header";
 
@@ -14,7 +14,7 @@ function EditPage() {
         <Body />
       </Container>
       <BookmarkModal />
-      <BookmarkGroupModal />
+      <GroupModal />
       <ConfirmModal />
     </>
   );
